backend/ai_engine.py
import google.generativeai as genai
import os
import json

# Placeholder for API Key - User should set this env variable or paste it here
# In a real app, use os.environ.get("GEMINI_API_KEY")
GEMINI_API_KEY = os.environ.get("GEMINI_API_KEY", "") 

def configure_gemini(api_key: str):
    """Configures the Gemini API with the provided key."""
    if not api_key:
        return False
    genai.configure(api_key=api_key)
    return True

def get_gemini_analysis(data_summary: str, api_key: str = None):
    """
    Sends a summary of the data to Gemini for quality analysis.
    """
    if not api_key and not GEMINI_API_KEY:
        return {"error": "API Key is missing"}
    
    key_to_use = api_key if api_key else GEMINI_API_KEY
    configure_gemini(key_to_use)
    
    model = genai.GenerativeModel('gemini-pro')
    
    prompt = f"""
    You are a Data Quality Expert. Analyze the following dataset summary and identify quality issues.
    
    Dataset Summary:
    {data_summary}
    
    Output JSON with the following structure:
    {{
        "quality_score": <0-100>,
        "issues": [
            {{ "column": "column_name", "issue_type": "type (e.g. Missing, Outlier, Format)", "description": "detail", "severity": "High/Medium/Low" }}
        ],
        "summary": "Brief text summary of data health"
    }}
    Return ONLY valid JSON.
    """
    
    try:
        response = model.generate_content(prompt)
        # cleanup json
        text = response.text.replace('```json', '').replace('```', '')
        return json.loads(text)
    except Exception as e:
        return {"error": str(e)}

def get_gemini_correction(row_data: dict, context: str, api_key: str = None):
    """
    Asks Gemini to correct a specific row based on context.
    """
    # Logic for individual row correction or batch correction...
    pass
